// pb/com/widedelivery/auth/proto/UpdateUserInput.ts
// Original file: protos/rpc_update_user.proto


export interface UpdateUserInput {
  'user_id'?: (string);
  'name'?: (string);
  'email'?: (string);
  'password'?: (string);
  'phone_number'?: (string);
  'photo'?: (string);
  'role'?: (string);
}

export interface UpdateUserInput__Output {
  'user_id': (string);
  'name': (string);
  'email': (string);
  'password': (string);
  'phone_number': (string);
  'photo': (string);
  'role': (string);
}